Example CF manifests for the app and stub
Includes a couple of CF manifests to get verify-frontend and the stub working.

In order to use them, you'll need to push the verify-stub manifest first, then update the verify-frontend manifest to point to the successfully pushed stub.

You may also need to change the name fields specified in the yaml files.

Notes:
- verify-frontend sets RAILS_ENV to development to enable /test-saml
- verify-frontend sets the commans because the Procfile uses 'front' and doesn't set the port from $PORT
- verify-frontend is currently configured to redirect logs to files, so they aren't available from cf logs
- verify-stub sets BUNDLE_WITHOUT to 'shenanigans' as otherwise the buildpack excludes development:test which is required